Fullscreen Clean-Up
Fixed full-screen stuff
Intro
Ran clean-up

+++ src/Core/zap/turrem/core/tech/list/TechList.java
@@ -36,7 +36,7 @@ public class TechList
 		Integer i = techIds.get(idetifier);
 		if (i != null)
 		{
-			return (int) i;
+			return i;
 		}
 		return -1;
 	}
